+++ src/FreeElephants/PsrRouter/FastRoute/DispatcherBuilder.php
@@ -5,6 +5,7 @@ namespace FreeElephants\PsrRouter\FastRoute;
 
 use FastRoute\Dispatcher;
 use FastRoute\RouteCollector;
+use FreeElephants\PsrRouter\OptionsHandlerInterface;
 use FreeElephants\PsrRouter\PathNormalization\Dummy;
 use FreeElephants\PsrRouter\PathNormalization\PathNormalizerInterface;
 use function FastRoute\simpleDispatcher;
@@ -13,10 +14,18 @@ class DispatcherBuilder
 {
     private array $routes = [];
     private PathNormalizerInterface $pathNormalizer;
+    private OptionsHandlerInterface $optionsHandlerPrototype;
 
-    public function __construct(PathNormalizerInterface $pathNormalizer = null)
+    public function __construct(
+        PathNormalizerInterface $pathNormalizer = null
+    )
     {
         $this->pathNormalizer = $pathNormalizer ?? new Dummy();
+    }
+
+    public function setOptionsHandlerPrototype(OptionsHandlerInterface $handler): self
+    {
+        $this->optionsHandlerPrototype = $handler;
     }
 
     public function addRoute(string $path, string $handler, string $method = 'GET'): self
@@ -39,12 +48,22 @@ class DispatcherBuilder
     {
         return simpleDispatcher(function (RouteCollector $routeCollector) {
             foreach ($this->routes as $path => $methodsOrHandler) {
-                if(is_array($methodsOrHandler)) {
+                $httpMethods = [];
+                if (is_array($methodsOrHandler)) {
                     foreach ($methodsOrHandler as $method => $handler) {
                         $routeCollector->addRoute($method, $path, $handler);
+                        $httpMethods[] = $method;
                     }
                 } else {
                     $routeCollector->addRoute('GET', $path, $methodsOrHandler);
+                }
+
+                if (isset($this->optionsHandlerPrototype) && !in_array('OPTIONS', $httpMethods)) {
+                    $allowedMethods = $httpMethods;
+                    $allowedMethods[] = 'OPTIONS';
+                    $optionsHandler = clone $this->optionsHandlerPrototype;
+                    $optionsHandler->setAllowedMethods($allowedMethods);
+                    $routeCollector->addRoute(['OPTIONS'], $path, $optionsHandler);
                 }
             }
         });
@@ -56,7 +75,7 @@ class DispatcherBuilder
         foreach ($config as $path => $methods) {
             $path = $this->pathNormalizer->normalizePath($path);
 
-            if(!is_array($methods)) {
+            if (!is_array($methods)) {
                 $methods = ['GET' => $methods];
             }
             $normalized[$path] = $methods;