+++ src/FreeElephants/PsrRouter/OptionsHandlerInterface.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace FreeElephants\PsrRouter;
+
+use Psr\Http\Server\RequestHandlerInterface;
+
+interface OptionsHandlerInterface extends RequestHandlerInterface
+{
+    public function setAllowedMethods(array $httpMethods): void;
+}